Minor fixes

+++ src/models/reports.rs
@@ -197,7 +197,7 @@ pub struct FlatAccount {
     pub category: String,
     pub normal_balance: String,
     pub is_contra: bool,
-    pub balance: f64,
+    pub balance: BigDecimal,
 }
 
 #[derive(Debug, Serialize, Deserialize, sqlx::FromRow)]
@@ -205,6 +205,6 @@ pub struct Node {
     pub code: String,
     pub name: String,
     pub category: String,
-    pub total: f64,
+    pub total: BigDecimal,
     pub children: Vec<Node>,
 }
